3.0 GenAI/3.1 Langchain/LangChain for LLM Application Development/RAG/requirements.txt
langchain==0.3.27
langchain-community==0.0.40
langchain-openai==0.1.7
langchain-core==0.3.72
chromadb==0.4.24
openai==1.30.1
python-dotenv==1.0.1
tiktoken==0.7.0
